20230611C++

+++ C++/20230612/b004.cpp
@@ -7,7 +7,7 @@ int main(){
 	double d, l, pi;
 	
 	while(cin >> d >> l){
-		float ans;
+		double ans;
 		pi = 2*acos(0);
         ans = pi * l/2 * sqrt((l/2)*(l/2) - (d/2)*(d/2));
         printf("%.3f\n", ans);
